Top post view navigation is now tag aware
The navigation controls in the post view above the image are now aware of any tags
that have been used to search. This way they navigate to other posts that match
the tag search.

The plus and minus buttons next to the tags in the list have also been added to the
post view so that the tag search can more easily be modified from that page.